+++ src/components/SearchDialog/MajorFilter.tsx
@@ -1,0 +1,69 @@
+import {
+  Box,
+  Checkbox,
+  CheckboxGroup,
+  FormControl,
+  FormLabel,
+  Stack,
+  Tag,
+  TagCloseButton,
+  TagLabel,
+  Wrap,
+} from "@chakra-ui/react";
+import { memo } from "react";
+
+export const MajorFilter = memo(
+  ({
+    allMajors,
+    selectedMajors,
+    onChange,
+  }: {
+    allMajors: string[];
+    selectedMajors: string[];
+    onChange: Function;
+  }) => {
+    return (
+      <FormControl>
+        <FormLabel>전공</FormLabel>
+        <CheckboxGroup
+          value={selectedMajors}
+          onChange={(values) => onChange("majors", values as string[])}
+        >
+          <Wrap spacing={1} mb={2}>
+            {selectedMajors.map((major) => (
+              <Tag key={major} size="sm" variant="outline" colorScheme="blue">
+                <TagLabel>{major.split("<p>").pop()}</TagLabel>
+                <TagCloseButton
+                  onClick={() =>
+                    onChange(
+                      "majors",
+                      selectedMajors.filter((v) => v !== major)
+                    )
+                  }
+                />
+              </Tag>
+            ))}
+          </Wrap>
+          <Stack
+            spacing={2}
+            overflowY="auto"
+            h="100px"
+            border="1px solid"
+            borderColor="gray.200"
+            borderRadius={5}
+            p={2}
+          >
+            {allMajors.map((major) => (
+              <Box key={major}>
+                <Checkbox key={major} size="sm" value={major}>
+                  {major.replace(/<p>/gi, " ")}
+                </Checkbox>
+              </Box>
+            ))}
+          </Stack>
+        </CheckboxGroup>
+      </FormControl>
+    );
+  }
+);
+MajorFilter.displayName = "MajorFilter";
