+++ src/lib/api.ts
@@ -1,22 +1,37 @@
 import axios from "axios";
 import { Lecture } from "../types";
+import { BASE_URL } from "../constants";
 
-let lecturesPromise: Promise<Lecture[]> | null = null;
+export const fetchMajors = () => axios.get<Lecture[]>(`${BASE_URL}/schedules-majors.json`);
+export const fetchLiberalArts = () =>
+  axios.get<Lecture[]>(`${BASE_URL}/schedules-liberal-arts.json`);
 
-const fetchMajors = () => axios.get<Lecture[]>("/schedules-majors.json");
-const fetchLiberalArts = () => axios.get<Lecture[]>("/schedules-liberal-arts.json");
+const apiCache = new Map<string, Promise<any>>();
 
-export const fetchAllLectures = async (): Promise<Lecture[]> => {
-  if (lecturesPromise) {
-    return lecturesPromise;
+const fetchWithCache = (key: string, fetcher: () => Promise<any>) => {
+  if (!apiCache.has(key)) {
+    console.log(`${key} API CALL`);
+    apiCache.set(
+      key,
+      fetcher().catch((error) => {
+        apiCache.delete(key);
+        throw error;
+      })
+    );
   }
+  return apiCache.get(key);
+};
 
-  lecturesPromise = Promise.all([fetchMajors(), fetchLiberalArts()]).then(
-    ([majorsResponse, liberalArtsResponse]) => {
-      const combinedLectures = [...majorsResponse.data, ...liberalArtsResponse.data];
-      return combinedLectures;
-    }
-  );
+export const fetchAllLectures = () => {
+  const majorsUrl = `${BASE_URL}/schedules-majors.json`;
+  const liberalArtsUrl = `${BASE_URL}/schedules-liberal-arts.json`;
 
-  return lecturesPromise;
+  return Promise.all([
+    fetchWithCache(majorsUrl, fetchMajors),
+    fetchWithCache(liberalArtsUrl, fetchLiberalArts),
+    fetchWithCache(majorsUrl, fetchMajors),
+    fetchWithCache(liberalArtsUrl, fetchLiberalArts),
+    fetchWithCache(majorsUrl, fetchMajors),
+    fetchWithCache(liberalArtsUrl, fetchLiberalArts),
+  ]);
 };
